+++ .gitignore
@@ -34,6 +34,7 @@ captures/
 
 # Intellij
 *.iml
+.idea/gradle.xml
 .idea/workspace.xml
 .idea/libraries
