+++ app/src/androidTest/java/com/oneweek/miluo/codingunbarred/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.oneweek.miluo.myapplication;
+package com.oneweek.miluo.codingunbarred;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
