+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.oneweek.miluo.myapplication">
+    package="com.oneweek.miluo.codingunbarred">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
@@ -11,8 +11,11 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
+        android:screenOrientation="sensorLandscape"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity
+            android:name=".MainActivity"
+            android:label="@string/main" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
@@ -20,17 +23,13 @@
         </activity>
         <activity
             android:name=".ResourceActivity"
-            android:label="@string/title_activity_resource"
-            android:parentActivityName=".MainActivity"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="MainActivity" />
-        </activity>
-        <activity android:name=".CourseSelectionActivity" />
-        <activity android:name=".ResourceSelectionActivity" />
-        <activity android:name=".DeveloperModeActivity" />
->>>>>>> Three other activities to navigate to
+            android:label="@string/resource" />
+        <activity android:name=".CourseSelectionActivity"
+            android:label="@string/course_selection" />
+        <activity android:name=".ResourceSelectionActivity"
+            android:label="@string/resource_selection" />
+        <activity android:name=".DevModeActivity"
+            android:label="@string/devmode" />
     </application>
 
 </manifest>
