+++ app/src/main/java/com/oneweek/miluo/codingunbarred/CourseSelectionActivity.java
@@ -1,13 +1,13 @@
-package com.oneweek.miluo.myapplication;
+package com.oneweek.miluo.codingunbarred;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class ResourceSelectionActivity extends AppCompatActivity {
+public class CourseSelectionActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_resource_selection);
+        setContentView(R.layout.activity_course_selection);
     }
 }
