+++ app/src/main/java/com/oneweek/miluo/codingunbarred/MainActivity.java
@@ -1,16 +1,8 @@
-package com.oneweek.miluo.myapplication;
+package com.oneweek.miluo.codingunbarred;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.widget.Button;
-import android.content.Intent;
-
-import android.view.View;
-import android.view.View;
-import android.content.Intent;
-import android.widget.Button;
-import android.content.Intent;
-
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 public class MainActivity extends AppCompatActivity {
@@ -34,7 +26,7 @@ public class MainActivity extends AppCompatActivity {
 
     // Sandbox button
     public void showSandbox(View view) {
-        Intent intent = new Intent(this, DeveloperModeActivity.class);
+        Intent intent = new Intent(this, DevModeActivity.class);
         startActivity(intent);
     }
 }
