+++ app/src/main/java/com/oneweek/miluo/codingunbarred/ResourceSelectionActivity.java
@@ -1,17 +1,13 @@
-package com.oneweek.miluo.myapplication;
-
-/**
- * Created by jakma on 7/25/2016.
- */
+package com.oneweek.miluo.codingunbarred;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
+public class ResourceSelectionActivity extends AppCompatActivity {
 
-public class DevModeActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_devmode);
+        setContentView(R.layout.activity_resource_selection);
     }
 }
