--- app/src/main/java/com/oneweek/miluo/myapplication/DeveloperModeActivity.java
@@ -1,13 +1,0 @@
-package com.oneweek.miluo.myapplication;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class DeveloperModeActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_developer_mode);
-    }
-}
