+++ app/src/main/res/layout/activity_course_selection.xml
@@ -7,6 +7,16 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.oneweek.miluo.myapplication.CourseSelectionActivity">
+    tools:context="com.oneweek.miluo.codingunbarred.CourseSelectionActivity"
+    android:background="#383838">
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceLarge"
+        android:text="@string/course_selection"
+        android:id="@+id/textView3"
+        android:layout_alignParentTop="true"
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="61dp" />
 </RelativeLayout>
