+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.oneweek.miluo.myapplication.MainActivity"
+    tools:context="com.oneweek.miluo.codingunbarred.MainActivity"
     android:background="#383838">
 
     <TextView
@@ -22,16 +22,17 @@
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Courses"
+        android:text="@string/courses"
         android:id="@+id/courses"
         android:background="#FFFFFF"
-        android:layout_below="@+id/textView"
+        android:layout_above="@+id/resources"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="40dp" />
+        android:layout_marginBottom="40dp"
+        android:onClick="showCourseSelection" />
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Resources"
+        android:text="@string/resources"
         android:id="@+id/resources"
         android:background="#FFFFFF"
         android:layout_centerVertical="true"
@@ -40,12 +41,12 @@
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Sandbox"
+        android:text="@string/sandbox"
         android:id="@+id/sandbox"
         android:background="#FFFFFF"
-        android:onClick="showSandbox"
         android:layout_below="@+id/resources"
-        android:layout_alignLeft="@+id/resources"
-        android:layout_alignStart="@+id/resources" />
+        android:layout_centerHorizontal="true"
+        android:layout_marginTop="40dp"
+        android:onClick="showSandbox" />
 
 </RelativeLayout>
