+++ app/src/main/res/menu/menu_resource.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.oneweek.miluo.myapplication.ResourceActivity">
+    tools:context=".ResourceActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
